# art_admin_backend/docs/swagger.yaml
basePath: /
definitions:
...
  request.BatchEmotionAnalysisRequest:
    properties:
      analysis_type:
        enum:
        - pattern
        - trend
        - insight
        type: string
      end_date:
        type: string
      start_date:
        type: string
      user_ids:
        items:
          type: integer
        type: array
    required:
    - analysis_type
    - end_date
    - start_date
    - user_ids
    type: object
...
  request.EmotionAnalysisRequest:
    properties:
      analysis_type:
        enum:
        - pattern
        - trend
        - insight
        type: string
      end_date:
        type: string
      start_date:
        type: string
    required:
    - analysis_type
    - end_date
    - start_date
    type: object
...
  response.PaginatedData:
    properties:
      current:
        description: 当前页码
        type: integer
      records:
        description: 记录列表
      size:
        description: 每页条数
        type: integer
      total:
        description: 总记录数
        type: integer
    type: object
  response.Response:
    properties:
      code:
        description: 状态码
        type: integer
      data:
        description: 数据
      msg:
        description: 消息
        type: string
    type: object
...
  service.EmotionAnalysisResult:
    properties:
      analysis_type:
        type: string
      analyzed_at:
        type: string
      data_points:
        type: integer
      end_date:
        type: string
      insights:
        items:
          type: string
        type: array
      patterns:
        items:
          $ref: '#/definitions/service.EmotionPattern'
        type: array
      recommendations:
        items:
          type: string
        type: array
      start_date:
        type: string
      summary:
        type: string
      trends:
        items:
          $ref: '#/definitions/service.EmotionTrend'
        type: array
      user_id:
        type: integer
    type: object
  service.EmotionPattern:
    properties:
      description:
        type: string
      frequency:
        type: integer
      percentage:
        type: number
      type:
        type: string
    type: object
  service.EmotionTrend:
    properties:
      change_rate:
        type: number
      confidence:
        type: number
      direction:
        description: '"improving", "declining", "stable"'
        type: string
      period:
        type: string
    type: object
host: localhost:48080
info:
  contact: {}
  description: Art Admin Backend API Documentation
  title: Art Admin API
  version: "1.0"
paths:
  /api/ai-analysis/{id}:
    delete:
      consumes:
      - application/json
      description: 删除特定的情绪分析记录
      parameters:
      - description: 分析ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除情绪分析记录
      tags:
      - AI分析
    get:
      consumes:
      - application/json
      description: 获取特定情绪分析的详细信息
      parameters:
      - description: 分析ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.EmotionAnalysisResult'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取情绪分析详情
      tags:
      - AI分析
  /api/ai-analysis/{id}/export:
    get:
      consumes:
      - application/json
      description: 导出情绪分析报告为PDF或JSON格式
      parameters:
      - description: 分析ID
        in: path
        name: id
        required: true
        type: integer
      - default: json
        description: 导出格式
        enum:
        - json
        - pdf
        in: query
        name: format
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 导出成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 导出情绪分析报告
      tags:
      - AI分析
  /api/ai-analysis/batch:
    post:
      consumes:
      - application/json
      description: 批量分析多个用户的情绪数据（管理员功能）
      parameters:
      - description: 批量分析请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.BatchEmotionAnalysisRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 分析成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 批量分析情绪
      tags:
      - AI分析
  /api/ai-analysis/emotions:
    post:
      consumes:
      - application/json
      description: 基于用户的情绪记录和日记数据进行AI情绪分析
      parameters:
      - description: 情绪分析请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.EmotionAnalysisRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 分析成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.EmotionAnalysisResult'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 分析用户情绪模式
      tags:
      - AI分析
  /api/ai-analysis/history:
    get:
      consumes:
      - application/json
      description: 获取用户的历史情绪分析记录
      parameters:
      - default: 1
        description: 当前页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页条数
        in: query
        name: pageSize
        type: integer
      - description: 分析类型
        in: query
        name: analysisType
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PaginatedData'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取情绪分析历史
      tags:
      - AI分析
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"